feat: 건강 bmi INT형으로 변경

+++ fitmily_backend/user-service/src/main/java/com/d208/fitmily/domain/health/dto/AddHealthRequestDto.java
@@ -25,7 +25,7 @@ public class AddHealthRequestDto {
     private Float weight;
 
     @Schema(hidden = true)
-    private Float bmi;
+    private Integer bmi;
 
     private List<String> OtherDiseases;
     private List<String> FiveMajorDiseases;
